Refactor code and add high level tests
